Put `${INPUT_RUBOCOP_FLAGS}` at the end of the other flags
Since PR #103 introduced the default --fail-level flag,
the severity level has become unchangeable due to this
modification.
Ideally, this flag (and others) should be overridable,
but the RuboCop command prioritizes the flags that are
specified last. This commit moves the user-specified
flags to the end of the hard-coded flags, allowing them
to be overridden.

Signed-off-by: moznion <moznion@mail.moznion.net>